Move error handling to new Error helper.
